Gestion quantité ++ quand joue de l'article sur la page produit + gestion suppression article dans le panier + tentative modification dynamique dans le panier + ajout order id et page confirmation

+++ front/js/cart.js
@@ -3,31 +3,36 @@ let produitsDansLocalStorage = JSON.parse(localStorage.getItem("produit"));
 var mainContainer = document.getElementById('cart__items');
 let affichagePanierDansHtml = []
 
-for (var i = 0; i < produitsDansLocalStorage.length; i++) {
+if (produitsDansLocalStorage === null || produitsDansLocalStorage == 0) {
+  mainContainer.innerHTML = '<p>Ajoutez des produits à votre panier !</p>';
+}
+else {
+  for (var i = 0; i < produitsDansLocalStorage.length; i++) {
 
-    affichagePanierDansHtml = affichagePanierDansHtml +
-        `<article class="cart__item">
-                <div class="cart__item__img">
-                  <img src="${produitsDansLocalStorage[i].imgProduit}" alt="Photographie d'un canapé">
-                </div>
-                <div class="cart__item__content">
-                  <div class="cart__item__content__description">
-                    <h2>${produitsDansLocalStorage[i].nomProduit}</h2>
-                    <p>${produitsDansLocalStorage[i].couleurProduit}</p>
-                    <p>${produitsDansLocalStorage[i].prixProduit}</p>
+      affichagePanierDansHtml = affichagePanierDansHtml +
+          `<article class="cart__item" data-id="${produitsDansLocalStorage[i].idProduit}">
+                  <div class="cart__item__img">
+                    <img src="${produitsDansLocalStorage[i].imgProduit}" alt="Photographie d'un canapé">
                   </div>
-                  <div class="cart__item__content__settings">
-                    <div class="cart__item__content__settings__quantity">
-                      <p>Qté : </p>
-                      <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${produitsDansLocalStorage[i].quantiteProduit}">
+                  <div class="cart__item__content">
+                    <div class="cart__item__content__description">
+                      <h2>${produitsDansLocalStorage[i].nomProduit}</h2>
+                      <p>${produitsDansLocalStorage[i].couleurProduit}</p>
+                      <p>${produitsDansLocalStorage[i].prixProduit}</p>
                     </div>
-                    <div class="cart__item__content__settings__delete">
-                      <p class="deleteItem">Supprimer</p>
+                    <div class="cart__item__content__settings">
+                      <div class="cart__item__content__settings__quantity">
+                        <p>Qté : </p>
+                        <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${produitsDansLocalStorage[i].quantiteProduit}">
+                      </div>
+                      <div class="cart__item__content__settings__delete">
+                        <p class="deleteItem">Supprimer</p>
+                      </div>
                     </div>
                   </div>
-                </div>
-        </article>`
-    mainContainer.innerHTML = affichagePanierDansHtml;
+          </article>`
+      mainContainer.innerHTML = affichagePanierDansHtml;
+  }
 }
 
 //Prix total et quantitée totale 
@@ -55,23 +60,30 @@ function getTotalPrice() {
 
 //Changer la quantité de manière dynamique dans le panier
 
-var quantityInputs = document.getElementsByClassName('itemQuantity');
-for (var i = 0; i < quantityInputs.length; i++) {
-  var input = quantityInputs[i];
-  input.addEventListener('change', quantityChanged);
-}
+function modifyQtt() {
+  let qttModif = document.querySelectorAll(".itemQuantity");
 
-function quantityChanged(event) {
-  var input = event.target
-  if (isNaN(input.value) || input.value <= 0) {
-    input.value = 1
+  for (let k = 0; k < qttModif.length; k++){
+      qttModif[k].addEventListener("change" , (event) => {
+          event.preventDefault();
+
+          //Selection de l'element à modifier en fonction de son id ET sa couleur
+          let quantityModif = produitsDansLocalStorage[k].quantiteProduit;
+          let qttModifValue = qttModif[k].valueAsNumber;
+          
+          const resultFind = produitsDansLocalStorage.find((el) => el.qttModifValue !== quantityModif);
+
+          resultFind.quantiteProduit = qttModifValue;
+          produitsDansLocalStorage[k].quantiteProduit = resultFind.quantiteProduit;
+
+          localStorage.setItem("produit", JSON.stringify(produitsDansLocalStorage));
+      
+          // refresh rapide
+          location.reload();
+      })
   }
-  localStorage.setItem("produit", JSON.stringify(produitsDansLocalStorage));
-  getNumberProduct();
-  getTotalPrice();
 }
-
-console.log(quantityInputs);
+modifyQtt();
 
 //Bouton supprimer du panier
 
@@ -83,21 +95,12 @@ for (let s = 0; s < boutonSupprimer.length; s++) {
         let selectionnerIdPourSuppression = produitsDansLocalStorage[s].idProduit;
         let selectionnerCouleurPourSuppression = produitsDansLocalStorage[s].couleurProduit;
 
-        produitsDansLocalStorage = produitsDansLocalStorage.filter(element => element.idProduit !== selectionnerIdPourSuppression && element.couleurProduit !== selectionnerCouleurPourSuppression);
+        produitsDansLocalStorage = produitsDansLocalStorage.filter(patate => patate.couleurProduit !== selectionnerCouleurPourSuppression || patate.idProduit !== selectionnerIdPourSuppression);
         localStorage.setItem("produit", JSON.stringify(produitsDansLocalStorage));
-
+        
+        boutonSupprimer[s].closest("article").remove();
     })
 }
-
-//let boutonSupprimer = document.querySelectorAll('.deleteItem');
-
-//for (let s = 0; s < boutonSupprimer.length; s++) {
-  //boutonSupprimer[s].addEventListener("click", function(event) {
-    //var boutonClick = event.target
-    //boutonClick.parentElement.parentElement.parentElement.parentElement.remove()
-
-  //})
-//}
 
 //Regex 
 
@@ -110,6 +113,7 @@ function ValidateEmail(mail)
     alert("You have entered an invalid email address!")
     return (false)
 }
+
 //Verification que le formulaire est correct avec regex 
 
 //Formulaire d'envoi 
@@ -122,21 +126,45 @@ let emailSelection = document.getElementById('email');
 
 
 let commander = document.getElementById('order')
-commander.addEventListener("click", () => {
-  
-  const contact = {
-    firstName: firstNameSelection.value,
-    lastName: lastNameSelection.value,
-    address: addressSelection.value,
-    city: citySelection.value,
-    email: emailSelection.value
-  }
+commander.addEventListener("click", (carotte) => {
 
-  const promise = fetch("http://localhost:3000/api", {
-  method: "POST",
-  body: JSON.stringify(contact),
-  headers: {
-    "Content-Type": "application/json",
+  carotte.preventDefault();
+
+  let idProduitFormulaire = [];
+  for (let i = 0; i < produitsDansLocalStorage.length; i++) {
+    idProduitFormulaire.push(produitsDansLocalStorage[i].idProduit);
   }
+  
+  const order = {
+    contact : {
+        firstName: firstNameSelection.value,
+        lastName: lastNameSelection.value,
+        address: addressSelection.value,
+        city: citySelection.value,
+        email: emailSelection.value,
+    },
+    products: idProduitFormulaire,
+} 
+
+const options = {
+    method: 'POST',
+    body: JSON.stringify(order),
+    headers: {
+        'Accept': 'application/json', 
+        "Content-Type": "application/json" 
+    },
+};
+
+fetch("http://localhost:3000/api/products/order", options)
+.then((response) => response.json())
+.then((data) => {
+    console.log(data);
+    localStorage.clear();
+    localStorage.setItem("orderId", data.orderId);
+
+    document.location.href = "confirmation.html";
 })
+.catch((err) => {
+    alert ("Problème avec fetch : " + err.message);
 });
+})
